app/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-rq==2.4.1
django-redis-cache==3.0.0
psycopg2==2.9.1
gunicorn==19.9.0